Bump SDK version to 35
Based on reading the upgrade guide we should not be affected. This will
not impact minimum version support but make sure we are compatible with
newer versions. This bump brings us in line with the sdl build. From my
testing this version seems to work, but it would be good to get a double
check on that.

Change-Id: Ia571023e1f501d56c3fa93f0fac0140bbff57a1b